third_party/blink/web_tests/virtual/popover-disabled/popover-disabled.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src='../../resources/testharness.js'></script>
<script src="../../resources/testharnessreport.js"></script>

<div id=elements>
  <div popover=auto>This content should be visible, if HTMLPopoverAttribute is disabled</div>
  <div popover=async>This content should be visible, if HTMLPopoverAttribute is disabled</div>
  <div popover=invalid>This content should be visible, if HTMLPopoverAttribute is disabled</div>
  <div popover="">This content should be visible, if HTMLPopoverAttribute is disabled</div>
  <div popover id=foo>This content should be visible, if HTMLPopoverAttribute is disabled</div>
</div>

<script>
window.onload = () => {
  if (window.testRunner)
    testRunner.dumpAsText();

  function popoverVisible(popover) {
    return !!(popover.offsetWidth || popover.offsetHeight || popover.getClientRects().length);
  }

  const elements = document.getElementById('elements').children;
  test(() => {
    assert_throws_dom("SyntaxError",() => {elements[0].matches(':open')},"The :open pseudo class shouldn't be available");
  },'Basic tests');

  for(let el of elements) {
    test(() => {
      assert_true(!!el);
      assert_true(popoverVisible(el));
      assert_equals(el.popover,undefined);
      assert_not_equals(window.getComputedStyle(el).display,'none');
    },`${el.outerHTML} should be stylable/visible and not display:none`);
  }

  function supportsPopover() {
    return HTMLElement.prototype.hasOwnProperty("popover");
  }

  test(() => {
    assert_false(supportsPopover());
  },`The popover IDL attribute should not be present on Element`);
};
</script>
