third_party/blink/web_tests/virtual/popover-disabled/popover-disabled-expected.txt

CONSOLE ERROR: Found a 'popover' attribute. If you are testing the popover API, you must enable Experimental Web Platform Features. If not, note that custom attributes must start with 'data-': https://html.spec.whatwg.org/multipage/dom.html#custom-data-attribute. This usage will *likely cause site breakage* when the popover API ships: https://chromestatus.com/feature/5463833265045504.
CONSOLE ERROR: Found a 'popover' attribute. If you are testing the popover API, you must enable Experimental Web Platform Features. If not, note that custom attributes must start with 'data-': https://html.spec.whatwg.org/multipage/dom.html#custom-data-attribute. This usage will *likely cause site breakage* when the popover API ships: https://chromestatus.com/feature/5463833265045504.
CONSOLE ERROR: Found a 'popover' attribute. If you are testing the popover API, you must enable Experimental Web Platform Features. If not, note that custom attributes must start with 'data-': https://html.spec.whatwg.org/multipage/dom.html#custom-data-attribute. This usage will *likely cause site breakage* when the popover API ships: https://chromestatus.com/feature/5463833265045504.
CONSOLE ERROR: Found a 'popover' attribute. If you are testing the popover API, you must enable Experimental Web Platform Features. If not, note that custom attributes must start with 'data-': https://html.spec.whatwg.org/multipage/dom.html#custom-data-attribute. This usage will *likely cause site breakage* when the popover API ships: https://chromestatus.com/feature/5463833265045504.
CONSOLE ERROR: Found a 'popover' attribute. If you are testing the popover API, you must enable Experimental Web Platform Features. If not, note that custom attributes must start with 'data-': https://html.spec.whatwg.org/multipage/dom.html#custom-data-attribute. This usage will *likely cause site breakage* when the popover API ships: https://chromestatus.com/feature/5463833265045504.
This is a testharness.js-based test.
PASS Basic tests
PASS <div popover="auto">This content should be visible, if HTMLPopoverAttribute is disabled</div> should be stylable/visible and not display:none
PASS <div popover="async">This content should be visible, if HTMLPopoverAttribute is disabled</div> should be stylable/visible and not display:none
PASS <div popover="invalid">This content should be visible, if HTMLPopoverAttribute is disabled</div> should be stylable/visible and not display:none
PASS <div popover="">This content should be visible, if HTMLPopoverAttribute is disabled</div> should be stylable/visible and not display:none
PASS <div popover="" id="foo">This content should be visible, if HTMLPopoverAttribute is disabled</div> should be stylable/visible and not display:none
PASS The popover IDL attribute should not be present on Element
Harness: the test ran to completion.
